LayoutTests/http/wpt/cross-origin-options/navigation-from-subframe-via-anchor-target.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Tests that 'Cross-Origin-Options: deny / allow-postmessage' prevents a cross-origin iframe from navigating us</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/utils.js"></script>
</head>
<body>
<script>

promise_test(t => {
    return withPopup("navigation-from-subframe-frame.py?value=deny&target=_top", false /* isCrossOrigin */).then((result) => {
        return new Promise((resolve) => {
            window.onmessage = t.unreached_func("Should not have navigated");
            t.step_timeout(() => {
               window.onmessage = null;
               resolve();
            }, 200);
        });
    });
}, "'Cross-Origin-Options: deny' prevents navigation from cross-origin sub-frame (using <a target=_top>)");

promise_test(t => {
    return withPopup("navigation-from-subframe-frame.py?value=allow-postmessage&target=_top", false /* isCrossOrigin */).then((result) => {
        return new Promise((resolve) => {
            window.onmessage = t.unreached_func("Should not have navigated");
            t.step_timeout(() => {
               window.onmessage = null;
               resolve();
            }, 200);
        });
    });
}, "'Cross-Origin-Options: allow-postmessage' prevents navigation from cross-origin sub-frame (using <a target=_top>)");

promise_test(t => {
    return withPopup("navigation-from-subframe-frame.py?value=allow&target=_top", false /* isCrossOrigin */).then((result) => {
        return new Promise((resolve) => {
            window.onmessage = () => {
                resolve();
            };
        });
    });
}, "'Cross-Origin-Options: allow' does not prevent navigation from cross-origin sub-frame (using <a target=_top>)");

promise_test(t => {
    return withPopup("navigation-from-subframe-frame.py?value=deny&target=_parent", false /* isCrossOrigin */).then((result) => {
        return new Promise((resolve) => {
            window.onmessage = t.unreached_func("Should not have navigated");
            t.step_timeout(() => {
               window.onmessage = null;
               resolve();
            }, 200);
        });
    });
}, "'Cross-Origin-Options: deny' prevents navigation from cross-origin sub-frame (using <a target=_parent>)");

promise_test(t => {
    return withPopup("navigation-from-subframe-frame.py?value=allow-postmessage&target=_parent", false /* isCrossOrigin */).then((result) => {
        return new Promise((resolve) => {
            window.onmessage = t.unreached_func("Should not have navigated");
            t.step_timeout(() => {
               window.onmessage = null;
               resolve();
            }, 200);
        });
    });
}, "'Cross-Origin-Options: allow-postmessage' prevents navigation from cross-origin sub-frame (using <a target=_parent>)");

promise_test(t => {
    return withPopup("navigation-from-subframe-frame.py?value=allow&target=_parent", false /* isCrossOrigin */).then((result) => {
        return new Promise((resolve) => {
            window.onmessage = () => {
                resolve();
            }; 
        });
    });
}, "'Cross-Origin-Options: allow' does not prevent navigation from cross-origin sub-frame (using <a target=_parent>)");

promise_test(t => {
    return withPopup("navigation-from-subframe-frame.py?value=deny&target=foo1", false /* isCrossOrigin */, "foo1").then((result) => {
        return new Promise((resolve) => {
            window.onmessage = t.unreached_func("Should not have navigated");
            t.step_timeout(() => {
               window.onmessage = null;
               resolve();
            }, 200);
        });
    }); 
}, "'Cross-Origin-Options: deny' prevents navigation from cross-origin sub-frame (using <a target=windowName)");

promise_test(t => {
    return withPopup("navigation-from-subframe-frame.py?value=allow-postmessage&target=foo2", false /* isCrossOrigin */, "foo2").then((result) => {
        return new Promise((resolve) => {
            window.onmessage = t.unreached_func("Should not have navigated");
            t.step_timeout(() => {
               window.onmessage = null;
               resolve();
            }, 200);
        });
    });
}, "'Cross-Origin-Options: allow-postmessage' prevents navigation from cross-origin sub-frame (using <a target=windowName)");

promise_test(t => {
    return withPopup("navigation-from-subframe-frame.py?value=allow&target=foo3", false /* isCrossOrigin */, "foo3").then((result) => {
        return new Promise((resolve) => {
            window.onmessage = () => {
                resolve();
            };
        });
    });
}, "'Cross-Origin-Options: allow' does not prevent navigation from cross-origin sub-frame (using <a target=windowName>)");

</script>
</body>
</html>
